UPDATE: Implementation of Product functionality

+++ src/modules/product/product.controller.ts
@@ -1,16 +1,20 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Patch,
+  Param,
+  Delete,
+  BadRequestException,
+} from '@nestjs/common';
 import { ProductService } from './product.service';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
 
-@Controller('product')
+@Controller('products')
 export class ProductController {
   constructor(private readonly productService: ProductService) {}
-
-  @Post()
-  create(@Body() createProductDto: CreateProductDto) {
-    return this.productService.create(createProductDto);
-  }
 
   @Get()
   findAll() {
@@ -19,16 +23,24 @@ export class ProductController {
 
   @Get(':id')
   findOne(@Param('id') id: string) {
+    if (!Number(id)) throw new BadRequestException('Icorrect ID format.');
     return this.productService.findOne(+id);
+  }
+
+  @Post()
+  create(@Body() createProductDto: CreateProductDto) {
+    return this.productService.create(createProductDto);
   }
 
   @Patch(':id')
   update(@Param('id') id: string, @Body() updateProductDto: UpdateProductDto) {
+    if (!Number(id)) throw new BadRequestException('Icorrect ID format.');
     return this.productService.update(+id, updateProductDto);
   }
 
   @Delete(':id')
   remove(@Param('id') id: string) {
+    if (!Number(id)) throw new BadRequestException('Icorrect ID format.');
     return this.productService.remove(+id);
   }
 }
